requirement.txt
asgiref==3.4.1
Django==3.2.8
psycopg2==2.9.1
python-dotenv==0.19.1
pytz==2021.3
sqlparse==0.4.2
